# swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "BuildTheEarth.net API",
    "version": "1.1"
  },
  "servers": [
    {
      "url": "https://api.buildtheearth.net/api/v1",
      "description": "Production instance"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development instance"
    }
  ],
...
  "paths": {
...
    "/applications/{id}": {
      "get": {
        "description": "Get one application",
        "parameters": [
          {
            "description": "The id of the application",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Extend BuildTeam Information",
            "in": "query",
            "name": "includeBuildteam",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Extend Reviewer Information",
            "in": "query",
            "name": "includeReviewer",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Extend Answer Information",
            "in": "query",
            "name": "includeAnswers",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": ["Applications"]
      },
      "post": {
        "description": "Review one application",
        "parameters": [
          {
            "description": "The id of the application",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Should the application be handeled as a trial application",
            "in": "query",
            "name": "isTrial",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": ["Applications"]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "openId": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://auth.buildtheearth.net/auth/realms/website/.well-known/openid-configuration"
      }
    },
...
  }
}